updating subscription service

+++ src/data4help/src/main/java/avila/schiatti/virdi/model/data/Location.java
@@ -4,22 +4,22 @@ import xyz.morphia.annotations.Embedded;
 
 @Embedded
 public class Location {
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
